refactor: centralize storage constants and improve code formatting
- Extract COURSE_MATERIALS_BUCKET to shared storage constants
- Fix TypeScript unused variable warnings with explicit typing
- Apply consistent code formatting across components
- Standardize imports and component structure
- Update Trigger.dev configuration with proper formatting
