+++ src/lib/constants/storage.ts
@@ -1,0 +1,5 @@
+/**
+ * Storage bucket constants
+ */
+
+export const COURSE_MATERIALS_BUCKET = "course-materials";
